requirements.txt
fastapi
uvicorn
python-dotenv
azure-ai-inference
azure-core
aiohttp
pydantic